requirements.txt
Pillow==5.2.0
tensorflow-gpu==1.10.0
typeguard==2.2.2
nibabel==2.3.0
scipy==1.1.0
